+++ lib/yoda/services.rb
@@ -4,6 +4,7 @@ module Yoda
     require 'yoda/services/current_node_explain'
     require 'yoda/services/comment_completion'
     require 'yoda/services/code_completion'
+    require 'yoda/services/loadable_path_resolver'
     require 'yoda/services/signature_discovery'
   end
 end