+++ spec/support/fixtures/lib/requires.rb
@@ -1,0 +1,2 @@
+require 'sample2'
+require 'set'